first struct for dog

+++ 0x0E-structures_typedef/dog.h
@@ -13,7 +13,7 @@
 struct dog
 {
 	char *name;
-	int age;
+	float age;
 	char *owner;
 };
